Added a property to shift the Z-axis of the Follow Cam/3D Cursor.
Its just a single float, but it works. I tried to add an XYZ VectorFloatProperty so you could adjust the X, Y, or Z position of the 3D cursor, but I couldn't figure out how to pass the Vector into the script to be used as a variable.

I left the VectorFloatProperty in the panel, since that part does work (its just commented out since it currently doesn't do anything).

If anyone can figure it out, feel free to modify the code.